componentized match tables + fixed pagination bug
