+++ resources/views/livewire/inc/subnav_game.blade.php
@@ -7,6 +7,9 @@
                 </h3>
                 <span class="badge text-bg-primary py-1 ms-3">CLASS {{ $game->category->class_label }}</span>
                 <span class="badge text-bg-secondary py-1 ms-2">{{ strtoupper($game->category->sex) }} CATEGORY</span>
+                @if($game->is_completed)
+                <span class="badge text-bg-success py-1 ms-2">COMPLETED</span>
+                @endif
             </div>
         </div>
         <hr class="mb-0">
@@ -18,23 +21,23 @@
                 <div class="d-flex flex-row align-items-center">
                     <ul class="navbar-nav">
                         <li class="nav-item">
-                            <a class="nav-link ps-1"
-                                wire:navigate href="{{ route('games.matches', ['event' => $event->id, 'game' => $game->id]) }}">Matches</a>
+                            <a class="nav-link ps-1" wire:navigate
+                                href="{{ route('games.matches', ['event' => $event->id, 'game' => $game->id]) }}">Matches</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link"
-                                wire:navigate href="{{ route('games.athletes', ['event' => $event->id, 'game' => $game->id]) }}">Athletes</a>
+                            <a class="nav-link" wire:navigate
+                                href="{{ route('games.athletes', ['event' => $event->id, 'game' => $game->id]) }}">Athletes</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link"
-                                wire:navigate href="{{ route('games.settings', ['event' => $event->id, 'game' => $game->id]) }}">Settings</a>
+                            <a class="nav-link" wire:navigate
+                                href="{{ route('games.settings', ['event' => $event->id, 'game' => $game->id]) }}">Settings</a>
                         </li>
                     </ul>
                 </div>
                 <div class="d-flex flex-row align-items-center">
                     <div class="pe-1 py-2">
-                        <a wire:navigate href="{{ route('events.show', ['event' => $event->id]) }}" class="custBtn custBtn-light"><i
-                                class="bi bi-arrow-left"></i>&nbsp Back to All Games</a>
+                        <a wire:navigate href="{{ route('events.show', ['event' => $event->id]) }}"
+                            class="custBtn custBtn-light"><i class="bi bi-arrow-left"></i>&nbsp Back to All Games</a>
                     </div>
                 </div>
             </div>
